extract bulk actions type
